io.home-assistant: address review comments
